--- delivery/src/main/java/takeoff/logisticsservice/msa/delivery/application/dto/request/GetDeliveryRequestDto.java
@@ -1,10 +1,0 @@
-package takeoff.logisticsservice.msa.delivery.application.dto.request;
-
-import java.util.UUID;
-
-public record GetDeliveryRequestDto(
-    UUID deliveryId,
-    UUID orderId
-) {
-
-}
