+++ delivery/src/main/java/takeoff/logisticsservice/msa/delivery/delivery/presentation/dto/request/PatchDeliveryRequestDto.java
@@ -1,0 +1,9 @@
+package takeoff.logisticsservice.msa.delivery.delivery.presentation.dto.request;
+
+import java.util.UUID;
+
+public record PatchDeliveryRequestDto(
+    UUID deliveryId,
+    String status) {
+
+}
