+++ delivery/src/main/java/takeoff/logisticsservice/msa/delivery/delivery/presentation/dto/request/PostDeliveryRequestDto.java
@@ -1,0 +1,7 @@
+package takeoff.logisticsservice.msa.delivery.delivery.presentation.dto.request;
+
+import java.util.UUID;
+
+public record PostDeliveryRequestDto(UUID orderID) {
+
+}
