+++ delivery/src/main/java/takeoff/logisticsservice/msa/delivery/deliveryRoute/presentation/dto/request/PostDeliveryRoutesRequest.java
@@ -1,0 +1,11 @@
+package takeoff.logisticsservice.msa.delivery.deliveryRoute.presentation.dto.request;
+
+import java.util.UUID;
+
+public record PostDeliveryRouteRequest(
+    UUID deliveryId,
+    UUID departureHubId,
+    UUID destinationHubId
+) {
+
+}
