--- delivery/src/main/java/takeoff/logisticsservice/msa/delivery/domain/entity/Delivery.java
@@ -1,74 +1,0 @@
-package takeoff.logisticsservice.msa.delivery.domain.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EmbeddedId;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.Table;
-import java.util.Objects;
-import java.util.UUID;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import takeoff.logistics_service.msa.common.domain.BaseEntity;
-
-@Entity
-@Table(name = "p_delivery")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Delivery extends BaseEntity {
-
-  @EmbeddedId
-  private DeliveryId id;
-
-  @Column(name = "order_id", nullable = false)
-  private UUID orderId;
-
-  @Column(name = "status", nullable = false)
-  @Enumerated(EnumType.STRING)
-  private DeliveryStatus status;
-
-  @Column(name = "from_hub_id", nullable = false)
-  private UUID fromHubId;
-
-  @Column(name = "to_hub_id", nullable = false)
-  private UUID toHubId;
-
-  @Builder
-  public Delivery(UUID orderId) {
-    this.orderId = orderId;
-    this.status = DeliveryStatus.ORDERED;
-
-    // TODO : 배송 경로 관련 로직 추가 + 허브 서비스와 연결
-    this.fromHubId = UUID.randomUUID();
-    this.toHubId = UUID.randomUUID();
-  }
-
-  public void modifyStatus(String status) {
-    switch (status) {
-      case DeliveryStatus.Status.ORDERED -> this.status = DeliveryStatus.ORDERED;
-      case DeliveryStatus.Status.DELIVERING -> this.status = DeliveryStatus.DELIVERING;
-      case DeliveryStatus.Status.COMPLETED -> this.status = DeliveryStatus.COMPLETED;
-      default -> throw new IllegalArgumentException("Invalid status: " + status);
-      // TODO : 글로벌 예외로 변경
-    }
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (!(o instanceof Delivery delivery)) {
-      return false;
-    }
-    return Objects.equals(id, delivery.id);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hashCode(id);
-  }
-}
