--- delivery/src/main/java/takeoff/logisticsservice/msa/delivery/presentation/dto/request/GetDeliveryRequest.java
@@ -1,17 +1,0 @@
-package takeoff.logisticsservice.msa.delivery.presentation.dto.request;
-
-import java.util.UUID;
-import takeoff.logisticsservice.msa.delivery.application.dto.request.GetDeliveryRequestDto;
-
-public record GetDeliveryRequest(
-    UUID deliveryId,
-    UUID orderId
-) {
-
-  public GetDeliveryRequestDto toApplicationDto() {
-    return new GetDeliveryRequestDto(
-        deliveryId,
-        orderId
-    );
-  }
-}
