--- delivery/src/main/java/takeoff/logisticsservice/msa/delivery/presentation/dto/request/PostDeliveryRequestDto.java
@@ -1,7 +1,0 @@
-package takeoff.logisticsservice.msa.delivery.presentation.dto.request;
-
-import java.util.UUID;
-
-public record PostDeliveryRequestDto(UUID orderID) {
-
-}
